refactor: AI Videos 초기 페이지 리팩터링 완료
