+++ src/components/TexttoVideo/TexttoVideo.css
@@ -1,0 +1,71 @@
+.texttovideo {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+.aiVideos__initial__btn {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  gap: 20px;
+  margin-bottom: 50px;
+}
+.aiVideos__initial__btn > button {
+  padding: 10px 15px;
+  font-size: 25px;
+  font-weight: bold;
+  color: #ffffff;
+  background-color: transparent;
+  border: none;
+  position: relative;
+  cursor: pointer;
+  transition: all 0.5s;
+}
+.aiVideos__initial__btn > button:hover {
+  transform: scale(1.05);
+}
+.aiVideos__initial__btn > button.active {
+  background: var(--gradient-text);
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  background-clip: text; /* 일부 브라우저 호환용 */
+  color: transparent; /* 혹시 모를 fallback */
+  border-bottom: 2px solid var(--gradient-text);
+}
+.aiVideos__initial__btn > button.active::after {
+  content: "";
+  position: absolute;
+  left: 0;
+  bottom: 0px;
+  width: 100%;
+  height: 2px;
+  background: var(--gradient-text);
+  border-radius: 2px;
+}
+
+.texttovideo__content {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 50px;
+}
+.texttovideo__generate {
+  width: calc(100% - 20px); /* 좌우 padding: 10px 가정 시 */
+  margin: 0 auto; /* 가운데 정렬 */
+  padding: 10px 0px;
+  border-radius: 15px;
+  font-size: 18px;
+  font-weight: bold;
+  transition: all 0.5s;
+  background: #d9d9d9;
+  cursor: not-allowed;
+}
+.texttovideo__generate.active {
+  background: var(--gradient-text);
+  cursor: pointer;
+}
+.texttovideo__generate.active:hover {
+  transform: scale(1.05);
+}
