+++ src/components/subcomponents/Information/Information.css
@@ -6,11 +6,19 @@
   height: 100%;
 }
 .information__screen__title {
+  display: flex;
+  align-items: center;
   font-size: 20px;
   font-weight: bold;
   color: #ffffff;
   margin-bottom: 10px;
   padding: 5px;
+  gap: 10px;
+}
+.information__screen__title > img {
+  width: 25px;
+  height: 25px;
+  filter: brightness(0) invert(1); /* 밝기 없애고 색 반전 => 흰색 */
 }
 .information__screen__main {
   padding: 10px;
