+++ src/components/subcomponents/Information/Information.jsx
@@ -1,10 +1,14 @@
 import React from "react";
 import "./Information.css";
+import infoIcon from "../../../assets/logo-information.png";
 
 const Information = () => {
   return (
     <div className="information__screen">
-      <header className="information__screen__title">Information</header>
+      <header className="information__screen__title">
+        <img src={infoIcon} alt="" />
+        Information
+      </header>
       <div className="information__screen__main">
         <span className="information__screen__subtitle1">상품이름</span>
         <textarea
