--- src/pages/AiVideos/ImagetoVideo/ImagetoVideo.css
@@ -1,50 +1,0 @@
-.imagetovideo {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.imagetovideo__options .gradient-box__inner {
-  background-color: #282e37 !important;
-}
-.imagetovideo__scrollable {
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  overflow-y: auto;
-  overflow-x: hidden;
-
-  /* ✅ 스크롤바 숨기기 */
-  scrollbar-width: none; /* Firefox */
-  -ms-overflow-style: none; /* IE and Edge */
-}
-
-.imagetovideo__scrollable::-webkit-scrollbar {
-  display: none; /* Chrome, Safari */
-}
-
-.imagetovideo__generate {
-  position: sticky;
-  bottom: 0; /* 부모 스크롤 영역 기준으로 하단에 고정 */
-  z-index: 10;
-  width: 97%;
-  margin-top: auto;
-  padding: 10px 0px;
-  border-radius: 15px;
-  font-size: 18px;
-  font-weight: bold;
-  transition: all 0.5s;
-  background: #d9d9d9;
-  cursor: not-allowed;
-  background-clip: padding-box; /* 테두리 겹침 방지용 */
-  backdrop-filter: blur(4px); /* 배경 흐림 효과 (선택사항) */
-}
-.imagetovideo__generate.active {
-  background: var(--gradient-text);
-  cursor: pointer;
-}
-.imagetovideo__generate.active:hover {
-  transform: scale(1.05);
-}
